refactoring in builders
